done successful connection between client&server

+++ feature/network/shared/src/lib.rs
@@ -1,2 +1,2 @@
 /// Порт сервера по умолчанию.
-pub const DEFAULT_PORT: u32 = 5000;
+pub const DEFAULT_PORT: u16 = 5000;
